Fixed 'restart microservice' info box
